fixed comments and build
