--- packages/core/constants/user-statuses.ts
@@ -1,5 +1,0 @@
-export enum UserStatuses {
-    ACTIVE = 'ACTIVE',
-    SUSPENDED = 'SUSPENDED',
-    PENDING_VERIFICATION = 'PENDING_VERIFICATION'
-}
