+++ packages/server/src/api/index.ts
@@ -4,6 +4,7 @@ const authLogin = `${auth}/login`;
 const authLoginFacebook = '/facebook';
 const authLoginGoogle = '/google';
 const authRegister = `${auth}/register`;
+const authConfirm = `${auth}/confirm`;
 const healthCheck = '/health-check';
 
 export const api = {
@@ -14,6 +15,7 @@ export const api = {
   healthCheck,
   authLoginFacebook,
   authLoginGoogle,
+  authConfirm,
 };
 
 export type Api = typeof api;
