+++ packages/server/src/constants/user-statuses.ts
@@ -1,0 +1,5 @@
+export enum UserStatuses {
+  ACTIVE = 'ACTIVE',
+  SUSPENDED = 'SUSPENDED',
+  PENDING_VERIFICATION = 'PENDING_VERIFICATION'
+}
