+++ packages/server/src/middlewares/validation.middleware.ts
@@ -4,7 +4,6 @@ import { RequestHandler } from 'express';
 import HttpException from '../exceptions/http.exception';
 
 export default function validationMiddleware<T>(
-  // eslint-disable-next-line @typescript-eslint/explicit-module-boundary-types
   type,
   skipMissingProperties = false
 ): RequestHandler {
