+++ packages/server/src/services/email-sending.service.ts
@@ -1,7 +1,8 @@
 import nodemailer from 'nodemailer';
 
-import { Pages } from 'core/constants/pages';
-import { QueryParams } from 'core/constants/query-params';
+import { Pages } from '../constants/pages';
+import { QueryParams } from '../constants/query-params';
+import {createActivationEmail} from "../utils/create-activation-email";
 
 export default class EmailSendingService {
   public async sendEmail(mailOptions: Record<string, unknown>): Promise<void> {
@@ -28,16 +29,12 @@ export default class EmailSendingService {
     }
   }
 
-  public async sendActivationMessage(userEmail: string, host: string, emailVerificationToken: string): Promise<void> {
+  public async sendActivationMessage(userEmail: string, name: string, host: string, emailVerificationToken: string): Promise<void> {
     const link = `${host}${Pages.EMAIL_CONFIRMATION_PAGE}?${QueryParams.EMAIL_VERIFICATION_TOKEN}=${emailVerificationToken}`;
     const mailOptions = {
       to: userEmail,
       subject: 'Please confirm your email account',
-      html: `
-        Hello,<br> Please Click on the link to verify your email.
-        Verification link expires in ${process.env.EMAIL_VERIFICATION_EXPIRATION_TIME} hours.
-        <br><a href="${link}">Click here to verify</a>
-      `
+      html: createActivationEmail(name, link)
     };
 
     await this.sendEmail(mailOptions);
